Remove packaging.py C headers-like pattern
Although __init__ appears to be a little less readible in practice, switch to pythonic pattern of using __init__ for all C header like references. In order to keep readability (and searchability) of the code, I have imported the symbols as {symbol}macro in order to let future coders know this is a macro and not a local variable.