Add feature support for skipping on secondary node
A feature request was made to allow these checks to skip based
on a puppet fact, shell var, or shell script to determine HA status
of your zabbix cluster.

Fix yaml eg, refactor logging, code readbility too

- Updated yaml template
- Added better logging, or cleaned up really large logging sequences (readability)
- Abstracted the sending of metrics to zabbix code, as it got kinda big (readability/oo)
- Abstracted the config syntax check of certain element details to config.py doesn't need to be part of checks. (readaibility)

Add feature to send check status with every check

Every zabbix item check needs a checkstatus item sent as a followup.
This will be executed after each check and have a customizable key format.
The input value will be either integer 1 or 0 depending on if the check has failed.
1 means failure.

Remove obsolete doublefork magic daemon pattern

As this is no longer suppose to (or expected) to avoid blocking under
the Zabbix daemon... we need to gut the doublefork code out.

This code is not needed or relevant to the new execution model (cron)

Autopep8 pass